Nguyen-dev

+++ AIstyle/__init__.py
@@ -15,16 +15,17 @@ password_postgreSQL = os.environ.get("password_postgreSQL")
 
 # print(SECRET_KEY, DB_NAME, user_postgreSQL, password_postgreSQL)
 
-def create_database(app):
+def create_database(app,remove):
     '''
     Chưa comment
     '''
     with app.app_context():
-        db.drop_all()
+        if remove == True:
+            db.drop_all()
         db.create_all()
     print("=> Created BD!")
 
-def create_app():
+def create_app(remove = False):
     '''
     Chưa comment
     '''
@@ -35,7 +36,7 @@ def create_app():
     app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
 
     db.init_app(app)
-    create_database(app)
+    create_database(app,remove)
 
     from AIstyle.user import user
     from AIstyle.home import home
